src/requirements.txt
# for python
pip
schedule
typer[all]

# for API
FastAPI
uvicorn[standard]

# for Monitoring
prometheus-api-client

# for Raspberry Pi
RPi.GPIO
